+++ src/collective/eeafaceted/collectionwidget/browser/views.py
@@ -3,6 +3,8 @@ from Products.CMFCore.utils import getToolByName
 from Products.Five.browser import BrowserView
 from eea.facetednavigation.browser.app.view import FacetedContainerView
 from eea.facetednavigation.subtypes.interfaces import IFacetedNavigable
+from collective.eeafaceted.collectionwidget.content.dashboardcollection import IDashboardCollection
+from collective.eeafaceted.collectionwidget.interfaces import NoCollectionWidgetDefinedException
 from collective.eeafaceted.collectionwidget.utils import getCollectionLinkCriterion
 
 
@@ -22,8 +24,10 @@ class RenderCategoryView(BrowserView):
 class RenderTermView(BrowserView):
 
     def display_number_of_items(self):
-        """If true, display the number of items in the collection."""
-        return True
+        """Display number of items in the collection."""
+        if not IDashboardCollection.providedBy(self.context):
+            return True
+        return self.context.showNumberOfItems
 
     def number_of_items(self):
         """Return the number of items in the collection."""
@@ -48,22 +52,27 @@ class FacetedDashboardView(FacetedContainerView):
 
     def __call__(self):
         criteria_holder = self._criteriaHolder
-        criterion = getCollectionLinkCriterion(criteria_holder)
-        # if we have the collection UID in the REQUEST, return self.index()
-        # so we avoid the portal_catalog search for collection
-        collectionUID = self.context.REQUEST.form.get('{0}[]'.format(criterion.__name__))
-        if collectionUID or not criterion.default:
-            return self.index()
-        if not self.request['HTTP_REFERER'].endswith('configure_faceted.html') and \
-           not self.request['URL'].endswith('folder_contents') and \
-           not self.request.get('no_redirect', '0') == '1':
-            catalog = getToolByName(self.context, 'portal_catalog')
-            brains = catalog(UID=criterion.default)
-            if brains:
-                collection = brains[0].getObject()
-                container = collection.aq_inner.aq_parent
-                if not container == criteria_holder and \
-                   IFacetedNavigable.providedBy(container):
-                    self.request.RESPONSE.redirect(container.absolute_url())
-                    return ''
+        criterion = None
+        try:
+            criterion = getCollectionLinkCriterion(criteria_holder)
+        except NoCollectionWidgetDefinedException:
+            pass
+        if criterion:
+            # if we have the collection UID in the REQUEST, return self.index()
+            # so we avoid the portal_catalog search for collection
+            collectionUID = self.context.REQUEST.form.get('{0}[]'.format(criterion.__name__))
+            if collectionUID or not criterion.default:
+                return self.index()
+            if not self.request['HTTP_REFERER'].endswith('configure_faceted.html') and \
+               not self.request['URL'].endswith('folder_contents') and \
+               not self.request.get('no_redirect', '0') == '1':
+                catalog = getToolByName(self.context, 'portal_catalog')
+                brains = catalog(UID=criterion.default)
+                if brains:
+                    collection = brains[0].getObject()
+                    container = collection.aq_inner.aq_parent
+                    if not container == criteria_holder and \
+                       IFacetedNavigable.providedBy(container):
+                        self.request.RESPONSE.redirect(container.absolute_url())
+                        return ''
         return self.index()