+++ src/collective/eeafaceted/collectionwidget/setuphandlers.py
@@ -9,3 +9,7 @@ def post_install(context):
     """Post install script"""
     if isNotCurrentProfile(context):
         return
+
+    # if plone.app.contenttypes is not installed, we need the BrowserLayer so various
+    # views like listing_view are available on DashboardCollection
+    context._tool.runImportStepFromProfile('profile-plone.app.contenttypes:default', 'browserlayer')