+++ src/collective/eeafaceted/collectionwidget/vocabulary.py
@@ -3,6 +3,8 @@
 from Products.CMFPlone.utils import safe_unicode
 from eea.facetednavigation.interfaces import ICriteria
 from eea.facetednavigation.interfaces import IFacetedNavigable
+from collective.behavior.talcondition.interfaces import ITALConditionable
+from collective.behavior.talcondition.utils import evaluateExpressionFor
 from collective.eeafaceted.collectionwidget.interfaces import ICollectionCategories
 from collective.eeafaceted.collectionwidget.widgets.widget import CollectionWidget
 from plone import api
@@ -22,12 +24,18 @@ class CollectionVocabulary(object):
         items = []
         current_url = context.absolute_url()
         for brain in self._brains(context):
+            collection = brain.getObject()
+            # if collection is ITALConditionable, evaluate the TAL condition
+            # except if current user is Manager
+            if ITALConditionable.providedBy(collection) and \
+               not evaluateExpressionFor(collection, extra_expr_ctx=self._extra_expr_ctx()):
+                    continue
+
             redirect_to = ''
             brain_folder_url = '/'.join(brain.getURL().split('/')[:-1])
             # if not in same folder and collection container is a faceted
             # we will redirect to this faceted to use criteria defined there
             if brain_folder_url != current_url or getRequest().get('force_redirect_to', False):
-                collection = brain.getObject()
                 collection_container = collection.aq_inner.aq_parent
                 if IFacetedNavigable.providedBy(collection_container):
                     # find the collection-link widget
@@ -37,7 +45,7 @@ class CollectionVocabulary(object):
                             redirect_to = self._compute_redirect_to(collection, criterion)
                             break
 
-            items.append(SimpleTerm(brain.getObject(),
+            items.append(SimpleTerm(collection,
                                     brain.UID,
                                     (safe_unicode(brain.Title),
                                      redirect_to)))
@@ -51,8 +59,9 @@ class CollectionVocabulary(object):
         catalog = api.portal.get_tool('portal_catalog')
         brains = catalog(
             path=dict(query='/'.join(root.getPhysicalPath())),
-            object_provides='plone.app.collection.interfaces.ICollection',
-            sort_on='getObjPositionInParent'
+            object_provides='plone.app.contenttypes.interfaces.ICollection',
+            enabled=True,
+            sort_on='getObjPositionInParent',
         )
         return brains
 
@@ -88,6 +97,11 @@ class CollectionVocabulary(object):
         query_url = '&'.join(default_criteria)
         return redirect_to.format(collection_container.absolute_url(),
                                   query_url)
+
+    def _extra_expr_ctx(self):
+        """To be overrided, this way, extra_expr_ctx is given to the
+           expression evaluated on the DashboardCollection."""
+        return {}
 
 
 CollectionVocabularyFactory = CollectionVocabulary()